+++ src/section.h
@@ -115,7 +115,7 @@ class SectionRefs
     const_iterator begin() const { return m_entries.cbegin(); }
     const_iterator end()   const { return m_entries.cend(); }
     bool empty() const { return m_entries.empty(); }
-    int size() const   { return m_entries.size();  }
+    int size() const   { return (int)m_entries.size(); }
 
   private:
     SectionInfoVec m_entries;
